refactored out queue logic
