+++ lib/index.js
@@ -1,12 +1,10 @@
 'use strict'
 
-var async = require('async')
 var chalk = require('chalk')
 var debug = require('debug-log')('mashape-analytics')
 var extend = require('xtend')
 var helpers = require('./helpers')
-var http = require('http')
-var pkg = require('../package.json')
+var Queue = require('./queue')
 var url = require('url')
 var util = require('util')
 
@@ -43,82 +41,8 @@ module.exports = function Agent (serviceToken, environment, options) {
     }
   }, options)
 
-  // setup event queue
-  // TODO specify worker pool
-  // TODO use msgpack + gzip?
-  this.queue = async.queue(function (entry, done) {
-    // append entry to log
-    self.alf.har.log.entries.push(entry)
-
-    // throttle
-    if (self.alf.har.log.entries.length < self.opts.queue.entries) {
-      debug('[%s] queued (%d/%d)', chalk.yellow('agent'), self.alf.har.log.entries.length, self.opts.queue.entries)
-
-      return done()
-    }
-
-    debug('[%s] sending (%d)', chalk.yellow('agent'), self.alf.har.log.entries.length)
-
-    // construct HTTP mesasge body
-    var postData = JSON.stringify(self.alf)
-
-    // immediatly reset entries object
-    self.alf.har.log.entries = []
-
-    // send the log
-    var request = http.request({
-      host: self.opts.host,
-      port: self.opts.port,
-      path: '/1.0.0/single',
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'User-Agent': util.format('%s/%s', self.alf.har.log.creator.name, self.alf.har.log.creator.version)
-      }
-    })
-
-    request.on('response', function (res) {
-      var chunks = []
-
-      res.on('data', function (chunk) {
-        chunks.push(chunk)
-      })
-
-      res.on('end', function () {
-        debug('[%s] %d %s: %s', chalk.magenta('socket'), res.statusCode, res.statusMessage, Buffer.concat(chunks))
-
-        done()
-      })
-    })
-
-    request.on('error', function (err) {
-      debug('[%s] problem with connection: %s', chalk.magenta('socket'), err.message)
-
-      done()
-    })
-
-    request.write(postData)
-    request.end()
-  })
-
-  // init HAR object
-  this.alf = {
-    version: '1.0.0',
-    serviceToken: serviceToken,
-    environment: environment || '',
-
-    har: {
-      log: {
-        version: '1.2',
-        creator: {
-          name: 'mashape-analytics-agent-node',
-          version: pkg.version
-        },
-
-        entries: []
-      }
-    }
-  }
+  // setup request queue
+  this.queue = new Queue(serviceToken, environment, self.opts)
 
   // API Recorder Middleware
   // TODO use tamper or tamper-esque method to get raw body
